Fixed Review comments
Renamed host_agent_test.go to label_flag_test.go
Renamed host_agent_integration_test.go to host_agent_test.go
