+++ agent/host_agent_test.go
@@ -1,64 +1,154 @@
 package main
 
 import (
+	"context"
+	"io/ioutil"
+	"os"
+	"os/exec"
+
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
+	"github.com/onsi/gomega/gbytes"
+	"github.com/onsi/gomega/gexec"
+	infrastructurev1alpha4 "github.com/vmware-tanzu/cluster-api-provider-byoh/apis/infrastructure/v1alpha4"
+	"github.com/vmware-tanzu/cluster-api-provider-byoh/common"
+	corev1 "k8s.io/api/core/v1"
+	"k8s.io/apimachinery/pkg/types"
 )
 
-var _ = Describe("Main", func() {
+var _ = Describe("Agent", func() {
 
-	Context("When the label flag is provided", func() {
+	Context("When the host is unable to register with the API server", func() {
 		var (
-			labels labelFlags
+			ns              *corev1.Namespace
+			err             error
+			hostName        string
+			fakedKubeConfig = "fake-kubeconfig-path"
+			session         *gexec.Session
 		)
+
 		BeforeEach(func() {
-			labels = make(labelFlags)
-		})
-		It("Should accept the single label flag", func() {
-			expLabels := labelFlags{"k1": "v1"}
-			Expect(labels.Set("k1=v1")).NotTo(HaveOccurred())
-			Expect(labels).Should(Equal(expLabels))
+			ns = common.NewNamespace("testns")
+			Expect(k8sClient.Create(context.TODO(), ns)).NotTo(HaveOccurred(), "failed to create test namespace")
+
+			hostName, err = os.Hostname()
+			Expect(err).NotTo(HaveOccurred())
 		})
 
-		It("Should accept the single label flag with comma separated kv pairs", func() {
-			expLabels := labelFlags{"k1": "v1", "k2": "v2"}
-			Expect(labels.Set("k1=v1,k2=v2")).NotTo(HaveOccurred())
-			Expect(labels).Should(Equal(expLabels))
+		AfterEach(func() {
+			err = k8sClient.Delete(context.TODO(), ns)
+			Expect(err).NotTo(HaveOccurred(), "failed to delete test namespace")
+			session.Terminate().Wait()
 		})
 
-		It("Should accept the single label flag with comma separated kv pairs with trailing comma", func() {
-			expLabels := labelFlags{"k1": "v1", "k2": "v2"}
-			Expect(labels.Set("k1=v1,k2=v2,")).NotTo(HaveOccurred())
-			Expect(labels).Should(Equal(expLabels))
+		It("should not error out if the host already exists", func() {
+			byoHost := &infrastructurev1alpha4.ByoHost{
+				TypeMeta: metav1.TypeMeta{
+					Kind:       "ByoHost",
+					APIVersion: "infrastructure.cluster.x-k8s.io/v1alpha4",
+				},
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      hostName,
+					Namespace: ns.Name,
+				},
+				Spec: infrastructurev1alpha4.ByoHostSpec{},
+			}
+			Expect(k8sClient.Create(context.TODO(), byoHost)).NotTo(HaveOccurred())
+
+			command := exec.Command(pathToHostAgentBinary, "--kubeconfig", kubeconfigFile.Name(), "--namespace", ns.Name)
+			session, err = gexec.Start(command, GinkgoWriter, GinkgoWriter)
+			Expect(err).NotTo(HaveOccurred())
+			Consistently(session).ShouldNot(gexec.Exit(0))
 		})
 
-		It("Should accept the multiple label flags", func() {
-			expLabels := labelFlags{"k1": "v1", "k2": "v2"}
-			Expect(labels.Set("k1=v1")).NotTo(HaveOccurred())
-			Expect(labels.Set("k2=v2")).NotTo(HaveOccurred())
-			Expect(labels).Should(Equal(expLabels))
+		It("should return an error when invalid kubeconfig is passed in", func() {
+			command := exec.Command(pathToHostAgentBinary, "--kubeconfig", fakedKubeConfig)
+			session, err = gexec.Start(command, GinkgoWriter, GinkgoWriter)
+			Expect(err).NotTo(HaveOccurred())
+			Eventually(session).Should(gexec.Exit(0))
+		})
+	})
+
+	Context("When the host agent is able to connect to API Server", func() {
+
+		var (
+			ns       *corev1.Namespace
+			session  *gexec.Session
+			err      error
+			workDir  string
+			hostName string
+		)
+
+		BeforeEach(func() {
+			ns = common.NewNamespace("testns")
+			Expect(k8sClient.Create(context.TODO(), ns)).NotTo(HaveOccurred(), "failed to create test namespace")
+
+			hostName, err = os.Hostname()
+			Expect(err).NotTo(HaveOccurred())
+
+			command := exec.Command(pathToHostAgentBinary, "--kubeconfig", kubeconfigFile.Name(), "--namespace", ns.Name, "--label", "site=apac")
+
+			session, err = gexec.Start(command, GinkgoWriter, GinkgoWriter)
+			Expect(err).NotTo(HaveOccurred())
+
+			workDir, err = ioutil.TempDir("", "host-agent-ut")
+			Expect(err).NotTo(HaveOccurred())
 		})
 
-		It("Should accept the multiple label flags with comma separated kv pairs", func() {
-			expLabels := labelFlags{"k1": "v1", "k2": "v2", "k3": "v3", "k4": "v4"}
-			Expect(labels.Set("k1=v1,k2=v2")).NotTo(HaveOccurred())
-			Expect(labels.Set("k3=v3,k4=v4")).NotTo(HaveOccurred())
-			Expect(labels).Should(Equal(expLabels))
+		AfterEach(func() {
+			err = k8sClient.Delete(context.TODO(), ns)
+			Expect(err).NotTo(HaveOccurred())
+			os.RemoveAll(workDir)
+			session.Terminate().Wait()
 		})
 
-		It("Should accept the multiple label flags with a mix of comma separated kv pairs and a single kv pair", func() {
-			expLabels := labelFlags{"k1": "v1", "k2": "v2", "k3": "v3"}
-			Expect(labels.Set("k1=v1,k2=v2")).NotTo(HaveOccurred())
-			Expect(labels.Set("k3=v3")).NotTo(HaveOccurred())
-			Expect(labels).Should(Equal(expLabels))
+		It("should register the BYOHost with the management cluster", func() {
+			byoHostLookupKey := types.NamespacedName{Name: hostName, Namespace: ns.Name}
+			createdByoHost := &infrastructurev1alpha4.ByoHost{}
+			Eventually(func() *infrastructurev1alpha4.ByoHost {
+				err := k8sClient.Get(context.TODO(), byoHostLookupKey, createdByoHost)
+				if err != nil {
+					return nil
+				}
+				return createdByoHost
+			}).ShouldNot(BeNil())
 		})
 
-		It("Should not accept the single label flag with only a key", func() {
-			Expect(labels.Set("k1")).To(MatchError("invalid argument value. expect key=value, got k1"))
+		It("should register the BYOHost with the passed labels", func() {
+			byoHostLookupKey := types.NamespacedName{Name: hostName, Namespace: ns.Name}
+			createdByoHost := &infrastructurev1alpha4.ByoHost{}
+			Eventually(func() map[string]string {
+				err := k8sClient.Get(context.TODO(), byoHostLookupKey, createdByoHost)
+				if err != nil {
+					return nil
+				}
+				return createdByoHost.ObjectMeta.Labels
+			}).Should(Equal(map[string]string{"site": "apac"}))
 		})
 
-		It("Should not accept the single label flag with comma separated errorneous input", func() {
-			Expect(labels.Set("k1=v1,k2")).To(MatchError("invalid argument value. expect key=value, got k1=v1,k2"))
+		It("should fetch networkstatus when register the BYOHost with the management cluster", func() {
+			byoHostLookupKey := types.NamespacedName{Name: hostName, Namespace: ns.Name}
+			Eventually(func() bool {
+				createdByoHost := &infrastructurev1alpha4.ByoHost{}
+				err := k8sClient.Get(context.TODO(), byoHostLookupKey, createdByoHost)
+				if err != nil {
+					return false
+				}
+				if len(createdByoHost.Status.Network) != 0 {
+					return true
+				}
+				return false
+			}).Should(BeTrue())
+
+		})
+
+		It("should only reconcile ByoHost resource that the agent created", func() {
+			byoHost := common.NewByoHost("random-second-host", ns.Name)
+			Expect(k8sClient.Create(context.TODO(), byoHost)).NotTo(HaveOccurred(), "failed to create byohost")
+
+			Consistently(session.Err, "10s").ShouldNot(gbytes.Say(byoHost.Name))
 		})
 	})
 })
